+++ adapters/wagmi-connector/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @blocto/wagmi-connector
 
+## 1.1.1-beta.2
+
+### Patch Changes
+
+- Fix evm switch chain is connected logic
+- Updated dependencies
+  - @blocto/sdk@0.5.5-beta.2
+
+## 1.1.1-beta.1
+
+### Patch Changes
+
+- Handle EVM switch chain if not connected.
+- Updated dependencies
+  - @blocto/sdk@0.5.5-beta.1
+
 ## 1.1.1-beta.0
 
 ### Patch Changes