+++ packages/blocto-sdk/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @blocto/sdk
 
+## 0.5.5-beta.2
+
+### Patch Changes
+
+- Fix evm switch chain is connected logic
+
+## 0.5.5-beta.1
+
+### Patch Changes
+
+- Handle EVM switch chain if not connected.
+
 ## 0.5.5-beta.0
 
 ### Patch Changes